requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-telegram-bot==20.7
psycopg2-binary==2.9.9
midtransclient==1.3.0
gunicorn==21.2.0
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
slowapi==0.1.9
httpx~=0.25.2
aiohttp
fastapi
gunicorn
httpx
midtransclient
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
python-telegram-bot
slowapi
uvicorn[standard]
